- Remove unnecessary wrappers throughout
