+++ tdrs-backend/tdpservice/etl/views.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from tdpservice.etl.exceptions import ActivePipelineRunError, PipelineValidationError
 from tdpservice.etl.models import ETLPipelineRun
 from tdpservice.etl.permissions import ETLPermissions
-from tdpservice.etl.runner import PipelineRunCreator
+from tdpservice.etl.runner import PipelineRunFactory
 from tdpservice.etl.serializers import (
     ETLPipelineRunCreateSerializer,
     ETLPipelineRunSerializer,
@@ -47,7 +47,7 @@ class PipelineRunViewSet(viewsets.ModelViewSet):
         serializer.is_valid(raise_exception=True)
 
         try:
-            pipeline_run = PipelineRunCreator.for_pipeline_key(
+            pipeline_run = PipelineRunFactory.for_pipeline_key(
                 serializer.validated_data["pipeline_key"]
             ).create(
                 parameters=serializer.validated_data["parameters"],
@@ -71,7 +71,7 @@ class PipelineRunViewSet(viewsets.ModelViewSet):
             raise ValidationError("Only failed ETL runs can be retried.")
 
         try:
-            pipeline_run = PipelineRunCreator.for_pipeline_key(
+            pipeline_run = PipelineRunFactory.for_pipeline_key(
                 failed_run.pipeline_key
             ).create(
                 parameters=failed_run.parameters,
